files/conversion/oxford-pets/requirements.txt
numpy==1.22.4
opencv-python==4.6.0.66
lxml
wai.common==0.0.40
wai.annotations.core>=0.1.8
Shapely==1.8.2
